backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
boto3>=1.39.8
botocore>=1.33.8
anthropic[bedrock]>=0.40.0
bcrypt==4.1.2
gunicorn==21.2.0
